refactor: move video I/O to minian.io (part of #309)
Colocate load/export paths in a dedicated io module per maintainer
direction on #309. Keep behavior unchanged; deprecate legacy helpers
and add utilities/visualization import shims for backward compatibility.
