+++ minian/test/test_pre_processing.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 import holoviews as hv
 
-from ..utilities import load_videos
+from ..io import load_videos
 from ..preprocessing import denoise, remove_background, stripe_correction
 
 dpath = "./demo_movies"
